CERT: plumbing and translation
