+++ libs/cardano-ledger-conformance/cardano-ledger-conformance.cabal
@@ -32,10 +32,12 @@ library
         Test.Cardano.Ledger.Conformance.SpecTranslate.Conway
         Test.Cardano.Ledger.Conformance.SpecTranslate.Conway.Base
         Test.Cardano.Ledger.Conformance.SpecTranslate.Conway.Pool
+        Test.Cardano.Ledger.Conformance.SpecTranslate.Conway.Cert
         Test.Cardano.Ledger.Conformance.ExecSpecRule.Core
         Test.Cardano.Ledger.Conformance.ExecSpecRule.Conway
         Test.Cardano.Ledger.Conformance.ExecSpecRule.Conway.Base
         Test.Cardano.Ledger.Conformance.ExecSpecRule.Conway.Pool
+        Test.Cardano.Ledger.Conformance.ExecSpecRule.Conway.Cert
 
     default-language: Haskell2010
     ghc-options:
