+++ libs/cardano-ledger-conformance/src/Test/Cardano/Ledger/Conformance/ExecSpecRule/Conway.hs
@@ -1,4 +1,5 @@
 module Test.Cardano.Ledger.Conformance.ExecSpecRule.Conway () where
 
 import Test.Cardano.Ledger.Conformance.ExecSpecRule.Conway.Base ()
+import Test.Cardano.Ledger.Conformance.ExecSpecRule.Conway.Cert ()
 import Test.Cardano.Ledger.Conformance.ExecSpecRule.Conway.Pool ()
