+++ libs/cardano-ledger-conformance/src/Test/Cardano/Ledger/Conformance/SpecTranslate/Conway/Pool.hs
@@ -1,6 +1,7 @@
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
+{-# LANGUAGE RecordWildCards #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE TypeOperators #-}
 {-# LANGUAGE UndecidableInstances #-}
@@ -8,11 +9,15 @@
 
 module Test.Cardano.Ledger.Conformance.SpecTranslate.Conway.Pool where
 
+import Cardano.Ledger.CertState
 import Cardano.Ledger.Core
+import Cardano.Ledger.Keys
+import Cardano.Ledger.PoolParams
 import Cardano.Ledger.Shelley.Rules
+import Data.Map.Strict (mapKeys)
 import qualified Lib as Agda
+import Test.Cardano.Ledger.Conformance
 import Test.Cardano.Ledger.Conformance.SpecTranslate.Conway.Base ()
-import Test.Cardano.Ledger.Conformance.SpecTranslate.Core
 import Test.Cardano.Ledger.Conway.TreeDiff
 
 instance
@@ -29,3 +34,23 @@ instance SpecTranslate ctx (ShelleyPoolPredFailure era) where
   type SpecRep (ShelleyPoolPredFailure era) = OpaqueErrorString
 
   toSpecRep e = pure . OpaqueErrorString . show $ toExpr e
+
+instance SpecTranslate ctx (PState era) where
+  type SpecRep (PState era) = Agda.PState
+
+  toSpecRep PState {..} =
+    Agda.MkPState
+      <$> toSpecRep (mapKeys (hashToInteger . unKeyHash) psStakePoolParams)
+      <*> toSpecRep (mapKeys (hashToInteger . unKeyHash) psRetiring)
+
+instance SpecTranslate ctx (PoolCert c) where
+  type SpecRep (PoolCert c) = Agda.TxCert
+
+  toSpecRep (RegPool p@PoolParams {ppId = KeyHash ppHash}) =
+    Agda.RegPool
+      <$> toSpecRep ppHash
+      <*> toSpecRep p
+  toSpecRep (RetirePool (KeyHash ppHash) e) =
+    Agda.RetirePool
+      <$> toSpecRep ppHash
+      <*> toSpecRep e
